Add CLI
